refactor: export run subcommand action to its own file
